Fixed regexes and miscount with vertical mean/sum.
The regexes allowed users to use cells such as A0 or B0 but these cells do not exist. This bug
has been fixed now.

The of a column of numbers was miss counting the number of elements. This has been fixed, the
counter starts at zero and counts the number of cells instead of working out the delta from
start and end cell.