Initial value is pulled without affecting other cells.
An object is pushed on the bus instead of a number. If the pusher_id is set to a cell id such B2, the onValue will only execute updateCell if it is cell B2. All other times the cell is updated.

Also Babel does not like ** for power, I have to use Math.pow(a,b). Babel thows a spurious bug, I found the cause by turning gulp tasks on and off.